refactor(store): implement re-export pattern with index files and update imports
