+++ src/pages/login-page/index.tsx
@@ -2,8 +2,8 @@ import { Link, Navigate } from 'react-router-dom';
 import { useState, useCallback, memo } from 'react';
 import { useAppDispatch, useAppSelector } from '../../hooks';
 import { AuthorizationStatus, CITIES } from '../../const';
-import { setCity } from '../../store/offers/offers-slice';
-import { selectAuthorizationStatus } from '../../store/user/user-selectors';
+import { setCity } from '../../store/offers';
+import { selectAuthorizationStatus } from '../../store/user';
 import { AppRoute } from '../../app/routes';
 import LoginForm from '../../components/login-form';
 import './login-page.css';
