+++ src/store/favorites/index.ts
@@ -1,0 +1,3 @@
+export * from './favorites-selectors';
+export * from './favorites-slice';
+export * from './favorites-thunks';
