+++ src/store/near-places/index.ts
@@ -1,0 +1,3 @@
+export * from './near-places-selectors';
+export * from './near-places-slice';
+export * from './near-places-thunks';
