+++ src/store/offer/index.ts
@@ -1,0 +1,3 @@
+export * from './offer-selectors';
+export * from './offer-slice';
+export * from './offer-thunks';
