+++ src/store/offers/index.ts
@@ -1,0 +1,3 @@
+export * from './offers-selectors';
+export * from './offers-slice';
+export * from './offers-thunks';
