+++ src/store/root-reducer.ts
@@ -1,11 +1,11 @@
 import { combineReducers } from '@reduxjs/toolkit';
 import { NameSpace } from '../const';
-import { userSlice } from './user/user-slice';
-import { offersSlice } from './offers/offers-slice';
-import { offerSlice } from './offer/offer-slice';
-import { favoritesSlice } from './favorites/favorites-slice';
-import { reviewsSlice } from './reviews/reviews-slice';
-import { nearPlacesSlice } from './near-places/near-places-slice';
+import { userSlice } from './user';
+import { offersSlice } from './offers';
+import { offerSlice } from './offer';
+import { favoritesSlice } from './favorites';
+import { reviewsSlice } from './reviews';
+import { nearPlacesSlice } from './near-places';
 
 export const rootReducer = combineReducers({
   [NameSpace.User]: userSlice.reducer,
